glib-2.0: add dependencies to fix ptest failures
There are a number of ptest failures due to executables not found in
the image. Since these executables are in FILES_${PN}-dev, depending
all of -dev which is an overkill. We take them out of -dev and put in
-utils, then depend on the latter instead.

When starting a glib ptest but decided to Ctrl-c interrupt midway,
./run-ptest cannot be restarted unless running `userdel glib2-test`
manually. Therefore adding a check prior to ptest will ensure the
ptest can be restarted.

(From OE-Core rev: 674602bf4eca88f090e14e0b528ae1751e50c71c)

Signed-off-by: Matthew Zeng <matthew.zeng@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>